KVM/Networking: Moving enter_kvm's SSH work to stage1/common/ssh for reuse purposes
